requirements.txt
GeoAlchemy2
geopandas
pandas
psycopg2
python-dotenv
requests
shapely
SQLAlchemy